fix: ResoultionBackground 컴포넌트 분리
